fix some C4244 warnings (mpc-hc 4724) fix C4706 warning (mpc-hc 4725) FLVSplitter.cpp был пропущен в r263 (mpc-hc 4703)
git-svn-id: svn://mpc-next.ru/mpc-be/trunk@275 d191f3aa-8897-45be-b9a4-56e5fb472b67

+++ src/filters/transform/MpaDecFilter/MpaDecFilter.cpp
@@ -1847,7 +1847,7 @@ HRESULT CMpaDecFilter::GetMediaType(int iPosition, CMediaType* pmt)
 	else if (subtype == MEDIASUBTYPE_Vorbis2) {
 		VORBISFORMAT2 *vf2 = (VORBISFORMAT2 *)mt.Format();
 		scmap_t* scmap = &m_scmap_default[vf2->Channels-1];
-		*pmt = CreateMediaType(GetSampleFormat(), vf2->SamplesPerSec, vf2->Channels, scmap->dwChannelMask);
+		*pmt = CreateMediaType(GetSampleFormat(), vf2->SamplesPerSec, (WORD)vf2->Channels, scmap->dwChannelMask);
 	}
 	else {
 		*pmt = CreateMediaType(GetSampleFormat(), wfe->nSamplesPerSec, min(2, wfe->nChannels));
